Correct managed mode.
Summary:
Corrects managed mode.
The standalone mcrouter is beeing spawned as a child process.
Every SIGTERM, SIGINT and SIGQUIT signals are beeing forwarded
from the parent to the child process. If the child fails to exit in a given
timeout (currently 3 seconds), a SIGKILL signal is then sent to it.

Test Plan:
Besides unit tests, the following manual tests were executed:

1) Send SIGTERM, SIGKILL, SIGQUIT and SIGINT to the child process
and watch it restart.
2) Send SIGTERM, SIGKILL, SIGQUIT and SIGINT to the parent process
and watch it forward the signal to the child, wait for it to die, and exit.
3) Send the same signals listed above to the parent, simulate a delay in
the child (in order for the timeout to expire) and watch the parent send a
SIGKILL to the child process and exit.

Reviewed By: pavlo@fb.com

Subscribers: hannesr, trunkagent, alikhtarov, ipozgaj

FB internal diff: D1713082

Tasks: 5665722

+++ mcrouter/test/MCProcess.py
@@ -81,6 +81,9 @@ class MCProcess(object):
         self.deletes = 0
         self.others = 0
 
+    def getprocess(self):
+        return self.proc
+
     def pause(self):
         if self.proc:
             self.proc.send_signal(signal.SIGSTOP)
@@ -129,7 +132,8 @@ class MCProcess(object):
 
         proc = self.proc
         if self.proc:
-            self.proc.terminate()
+            if self.proc.returncode is None:
+                self.proc.terminate()
             self.proc.wait()
             self.proc = None
 
@@ -382,6 +386,9 @@ class MCProcess(object):
         self.socket.sendall("version\r\n")
         return self.fd.readline()
 
+    def shutdown(self):
+        self.socket.sendall("shutdown\r\n")
+        return self.fd.readline()
 
 def sub_port(s, substitute_ports, port_map):
     parts = s.split(':')
